This project may be professional, but it's not big
I've decided to remove the code of conduct.  While this is a
professional project, i do not feel as though it is big enough for a
large code of law like the one that i had to make sense.  This coincides
with me simplifying the labels available on Github for issues and pull
requests.  Maybe i'll end up using somethings like these in the
future for bigger projects.

I also fixed a couple of formatting issues.