Added to manpage, fixed makefile
The manpage is now a little more descriptive about what the program will
actually do, including how it provides facilities for understanding the
origin of a particular splash file.  It's now more consistent between
the library and the program about the output, too.

The makefile has been changed based on how `make` is documented by
POSIX.  It now feels like it shouldn't work.