Added some test splash files
I plan to port into this program some splashes that i've collected for
other programs later.  It's just going to take a hell of a long time
because there are a lot of them, and i would rather get the code done
first.  So these files are just here for testing and won't appear
otherwise.

Also, i apparently changed the manpage.