Fetching
This is some of the most disasterous code i have ever written.  Anyways,
next up is randnum psuedofiles, then processing the retrieved string for
escape sequences, then the header (which i forgot about at a certain
point in the code in this commit…oops), then actually getting the offsets
and lengths of each line, then accounting for weights.  Shouldn't be too
hard—that's only five things.