Weights accounted for
This is some stupid code i've just written, but at least it was easy to
write this time.  And hooray for implicit typecastingeugh bloody balls
that hurt as it came out of my tips.

Anyway, next up is bugfixing, and then i'll add in the actual splash
files (obviously that will coincide with removing the placeholders).